+++ cogs/markov.py
@@ -48,10 +48,10 @@ class Markov(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message):
+        if message.author.bot or not message.content or message.guild is None:
+            return
         if not os.path.exists(f"usr/markov//{message.guild.id}/"):
             os.mkdir(f"usr/markov//{message.guild.id}/")
-        if message.author.bot or not message.content:
-            return
         if f"<@{self.bot.user.id}>" in message.content:
             if f"<@{self.bot.user.id}> ^" in message.content:
                 seed = message.content.replace(f"<@{self.bot.user.id}> ^", "")
